bookstore/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.0.4
psycopg2==2.8.5
whitenoise==5.1.0
dj-database-url==0.5.0
requests==2.23.0